Make Python interruptible on Windows (#10282)
This change addresses a long-standing issue that made it impossible to
interrupt Python on Windows. It is really just the last mile of the
change; most of the code that makes this work was added to the kernel
supervisor several months ago.

Interrupting ipykernel on Windows is (even before this change) a
byzantine process that involves creating an event object, passing its
handle to the kernel in a specially named `JPY_INTERRUPT_EVENT` object,
and then signaling the object from the controlling process to trigger an
interrupt.

Things are further complicated by the way that we specifically use
ipykernel: because ipykernel is not running on the main thread, it
doesn't pick up the event signal.

The fix is to create a top-level loop that listens for the event on
Windows, and to inject a `KeyboardInterrupt` when we get it.

Addresses #4604. 


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fixed interrupt of Python console on Windows (#4604)


### QA Notes

Note that not all Python statements are interruptible, and in particular
that `time.sleep(...)` doesn't have any way to check for interrupts so
isn't interruptible in Positron. In general, this change should bring us
to parity with other ipykernel frontends like VS Code.

To avoid a busy loop we only poll for interrupts every 200ms or so, so
also note that interrupts won't be instantaneous.

A new test has been added to the e2e suite that checks the new behavior.

---------

Signed-off-by: Jonathan <jonathan@posit.co>
Co-authored-by: Wasim Lorgat <mwlorgat@gmail.com>