Positron notebooks: Refactor to use existing menu and actions systems (#10276)
This PR refactors Positron notebooks to use builtin menu and action
systems, and adds a few commands to the command palette (we can
add/remove as needed via the `f1` property), addresses
https://github.com/posit-dev/positron/issues/10275.

This will make it easier to integrate with more of the existing Code OSS
functionality (like enabling the command palette above, and probably
much more) as well as to reuse other existing React components built for
these common types (for example, see the new extracted `Icon`
component).

I initially discovered that this was possible while working on #10248,
when I couldn't figure out how else to set a `precondition` other than
to use `Action2`.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- N/A

### QA Notes

This is quite a broad change, so we should kick the tires and make sure
that nothing broke. I used Claude pretty heavily for this one given the
amount of tedious changes, so more reason to double-check!

@:positron-notebooks