Positron notebooks: Connect to existing notebook context keys & restart after shutdown (#10297)
This PR:

1. Connects Positron notebooks to the existing notebook context keys
that are used elsewhere in core Positron (e.g. in point 3 below) and in
extensions (#10296).
2. Updated the way we handle `INotebookEditor` so that we can implement
different subsets of the interface for different features, we now use
this for (a) extension API, and (b) notebook context keys.
3. Enables the _Restart Kernel_ command for any notebook with one of our
kernels selected, even when the session is shutdown.
4. Restarting from an exited session now starts a new session (#10227).

https://github.com/user-attachments/assets/e13c8cbf-b1a3-4c1e-917f-57b269e8045d

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- N/A

### QA Notes

See the repro in #10227 and the attached video.

@:positron-notebooks