Add configuration for top action bar visibility (#10328)
Addresses https://github.com/posit-dev/positron/issues/6479

This PR adds a new setting `workbench.topActionBar.visible` that works
very similarly to `workbench.statusBar.visible`, including maintaining a
bidirectional sync with the UI and setting. If you click here in the UI:

<img width="629" height="452" alt="Screenshot_2025-11-02_at_3_39_32 PM"
src="https://github.com/user-attachments/assets/3b17304c-c2d5-4645-867b-addb05e282fb"
/>

the setting will update in your user `settings.json`, just like how the
`workbench.statusBar.visible` setting behaves. I think these two are a
good comparison, because they both contain a lot of useful information,
but folks may prefer to set up a more streamlined UI depending on their
preferences.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Added a new `workbench.topActionBar.visible` setting to control
whether the Positron top bar is visible

#### Bug Fixes

- N/A


### QA Notes

- Open up your user `settings.json` so you can see it in the editor
- Open up the "Customize Layout" quickpick
- Toggle the status bar off (notice your `settings.json`) and then back
on; this is existing behavior that we already have
- Now toggle the top bar off (notice your `settings.json`) and then back
on; this is the new behavior that this PR adds and it should work the
same
- Try putting both in your `settings.json` again and closing Positron;
when you open it up again, the settings should be respected