e2e: add screenshot for render markdown test (mac & rhel) (#10391)
### Summary
* Adding the expected markdown screenshot for the mac run. All other
platforms already have their expected screenshots. Also updated to NOT
assert for now since there are UI rendering changes expected for
notebooks. Will just collect screenshots and attach to report.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

@:positron-notebooks @:web @:win