catalog explorer: bundle `snowflake-sdk` (#10413)
Similar to `positron-duckdb`, we need to include the `node_modules` to
correctly resolve types for snowflake connections.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- #10104 Bundles snowflake packages needed to run in release builds.


### QA Notes

Snowflake should run on a released build, I tested this locally by
running

```
npm run gulp vscode
```

and confirmed the catalog explorer worked with snowflake.