Use normal whitespace in activity html outputs (#10412)
Fixes an issue causing Pandas data frames to render very unpleasantly in
the Console.

The issue was that the `white-space` property for the Console was set
with the intent of preserving whitespace formatting for console text,
but it was ALSO preserving whitespace formatting inside the sanitized
HTML components.

The fix is to use `normal` white space rendering inside HTML outputs. 

Addresses https://github.com/posit-dev/positron/issues/10382. 

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fix rendering of Python data frames in the Console (#10382)

Test tags: `@:console`