Assistant: pare back R and Python examples (#10437)
Addresses #10435.

In conversing with PA on `main` vs. in this PR, I don't observe any
differences in behavior with Claude Sonnet 4.5.

Token counts of the `<examples-lang>` sections are now drastically
lower:

|        | Before | After |
|--------|--------|-------|
| R      |     1960   |    240   |
| Python |    2310    |    620 (first-go was 427)   |

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes
@:assistant

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->

* I don't have as much of an eye for the "feel" with Python. I can
observe that it uses the packages noted in the prompt seemingly as
fluently as before, but I don't know enough to know whether it's using
no-longer-preferred APIs.

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->