Add support for popping entire plots pane to an auxiliary window (#10329)
Addresses https://github.com/posit-dev/positron/issues/7054

I was playing around with the existing _Chat: New Chat Window_ command
and it made me curious how difficult it would be to wire up the plots
pane functionality in the same way. The main architectural thing here is
that the plots UI would then need to render in two pretty different
container types, a view pane (where it currently lives) and an editor
pane (which provides a whole bunch of existing functionality for
auxiliary windows). To do this, I (with Claude Code's help) changed from
using `PositronPlotsViewPane` to plain old `IReactComponentContainer`.
The `PositronPlots` component was **already only using** properties from
`IReactComponentContainer`:
- `width`, `height`, `containerVisible`
- `onSizeChanged`, `onVisibilityChanged`, `onPositionChanged`
- `takeFocus()`

It was **not** using any `PositronPlotsViewPane`-specific functionality
like:
- View pane lifecycle methods
- Panel rendering internals
- ViewDescriptor properties

I _think_ this is a good way to go, so that we can render the plots UI
in both of these types of places?

### Some questions:

- Right now, all the plots content is mirrored in both places. Do we
like this? Or do we want to move it all to the auxiliary window, and
then back?
- Right now, all the plot actions are the same in both places. Do we
like this? Do any of them not make sense in the auxiliary window (such
as "Open in new window")?
- Do we like "Plots Gallery" as a name for this?
- Right now, this is _only_ available as a command from the command
palette. I think I'd like to go with that at first, before surfacing
this on the UI after we get some use on it. Thoughts?

@:plots

### Release Notes

#### New Features

- Added new command _Plots: Open Plots Gallery in New Window_ to open
the whole Plots pane UI in an auxiliary window. You can also pop out the
new standalone window from the plots action bar.
- Added new setting `plots.historyPolicy` to control the when the plot
history filmstrip is visible.

#### Bug Fixes

- N/A


### QA Notes

- Make a handful of plots, with both R and Python
- Run the new command _Plots: Open Plots Gallery in New Window_ (or
click the new button) to pop out a duplicate of the whole Plots pane UI
into an auxiliary window
- Check that all the same actions on the plots work, as they do when in
the main window
- The Plots pane should pop back to the default location when you close
the auxiliary window
- The new setting `plots.historyPolicy` now controls the behavior of the
plot history filmstrip

It doesn't look to me like we currently have any way to do automated
tests any auxiliary window functionality; this might be something for us
to put on the backlog.