Assistant: simplify `getProjectTree`'s exclusion arguments (#10425)
Addresses #10006.

This PR aims to simplify PA's interface to searching the projectTree.
Before, the relevant arguments to the tool were:

* `include`: Globs to include
* `exclude`: Globs to exclude
* `replaceDefaultExcludes`: Whether to skip the default excludes in this
tool's metadata.
* `ignoreFiles`: Whether to respect `.gitignore` rules
* `excludeSettings`: Which of the VS Code `settings.json` settings to
respect
* `filterResults`: Whether to apply the above three exclusion rules.

With this PR, the interface is:

* `include`: Globs to include
* `exclude`: Globs to exclude
* `skipDefaultExcludes`: Whether to skip the default excludes in this
tool's metadata, the `searchAndFilesExclude` metadata, and `.gitignore`.
As in "bypass the default exclusions—I need everything."

As a way to evaluate how well the model can interface with the exclusion
rules, I used the prompt "What files are in `extensions/node_modules`?"
in the Positron workspace. The "default" behavior seemed to be searching
with that `includes` and then, upon seeing the empty result, concluding
that there's nothing there.

<img width="1116" height="936" alt="model_needs_to_see_exclusions"
src="https://github.com/user-attachments/assets/1d61a7d5-16f9-41d4-8b2c-6d24676139c9"
/>

For a while, I tried to prompt my way into the model recognizing that
relevant exclusions were applied, incl. to the point of listing them all
out in the package.json:

```
"skipDefaultExcludes": {
  "type": "boolean",
    "description": "Whether to skip all automatic exclusions (.gitignore rules, VS Code exclude settings, and 
    default exclusions). When true, only the explicit `exclude` patterns will be applied. The default excludes are: 
    **/.build/**, **/.git/**, **/.devcontainer/**, **/.hg/**, **/.ipynb_checkpoints/**, **/.pytest_cache/**, **/.svn/**, 
    **/.venv/**, **/.Rproj.user/**, **/.vscode/**, **/__pycache__/**, **/dist/**, **/node_modules/**, **/renv/**, 
    **/venv/**, .DS_Store, .RData, .Rhistory, desktop.ini, package-lock.json, yarn.lock, Thumbs.db, **/*.a, 
    **/*.bak, **/*.csv~, **/*.js.map, **/*.log, **/*.o, **/*.pyo, **/*.so, **/*.tmp.",
  "default": false
},
```

What I ended up opting for was to include a note in the tool result that
files were excluded from the search when that is the case (and results
are sparse):

```
2025-11-06 10:30:25.645 [debug] [tool] Tool getProjectTree returned result: [
  {
    "$mid": 21,
    "value": ""
  },
  {
    "$mid": 21,
    "value": "331 results were excluded. Set `skipDefaultExcludes` to `true` to see them."
  }
]
```

This is implemented by running a second search without the default
exclusions when `skipDefaultExcludes` is `false` (notably, the default)
and results are sparse (effective max / 10, so <5 results by default). I
was initially pretty concerned about the performance implications here,
but found that the elapsed time to call this tool (even across the
entire Positron repository) was still _very_ snappy compared to the
generation of tokens from the model.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->

* To test the worse-case performance/latency of the tool, open the
Positron workspace and ask "Search for files with `positron` in their
name with getProjectTree" The latency of the tool return should be
similar to what it was before.
* To test that the agent can find files excluded by default when it
needs to, open the Positron workspace and ask "What files are in
`extensions/node_modules`?" The agent should first see no results but
also that some were excluded, then either attempt to call the tool again
with `skipDefaultExcludes` to `true` or just note that many matches were
excluded.

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

---------

Signed-off-by: Simon P. Couch <simonpatrickcouch@gmail.com>
Co-authored-by: sharon <sharon-wang@users.noreply.github.com>