Add positron notebook awareness to assistant.  (#10440)
This PR adds awareness of positron notebooks to the assistant chat pane.

@:assistant @:positron-notebooks

Addresses #613, #8555, #9142, #9431, #10342, #7373, #9662, #10387,
#10357, #10388 and #10356

When a positron notebook editor is open and active and it's associated
ipynb is attached as context to the chat then we send in context about
the notebook (cell count, language, selected cells, etc).

For a much more thorough overview see the new doc about it
`extensions/positron-assistant/src/tools/positron-assistant-notebook-mode.md`

<img width="900" height="941" alt="image"
src="https://github.com/user-attachments/assets/1b9dae37-fb00-46be-9b85-03d7346d658f"
/>

In addition we supply a few tools for the notebook to manipulate
notebooks (running, editing, deleting and looking at cell output).

This is a rough first step but is intended to provide the backbone upon
which we build and polish.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

When the chat is open and a positron notebook is open next to it the
chat should know about the notebook itself.

Prompt:
> What can you tell me about this notebook?

Should return useful info.

When the notebook is not attached in context, then the assistant should
not get any of this info. Test this by removing from context and asking
the same question.
<img width="981" height="554" alt="image"
src="https://github.com/user-attachments/assets/6ee61e98-33e7-44d7-9257-e5c7a52afe6f"
/>

---------

Signed-off-by: Nick Strayer <nick.strayer@posit.co>
Co-authored-by: George Stagg <george.stagg@posit.co>