Assistant: clarify relative vs. absolute paths in tool descriptions (#10521)
Addresses #8849. These feel to me like the smoking gun of that issue.

Note that both of those tools use `getUriForFileOpenOrInsideWorkspace`
and can handle relative paths just fine!

My reflex is that we ought to now remove `filepaths.md`, as the behavior
described there should otherwise be the "default" of models like Claude
Sonnet 4.5 and that file contains more prompting that needs to be
maintained. Fine with me to wait on that until these changes have had
some time in front of users without the noted issue appearing any
longer, though. :)

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

cc @jonvanausdeln for an eval case—when using both of these tools, PA
should use relative paths and the provided relative paths shouldn't
cause any issues with reads or edits:
 
<img width="1653" height="915" alt="Screenshot 2025-11-12 at 4 18 00 PM"
src="https://github.com/user-attachments/assets/bd546ffb-598b-483e-8bba-70b4345b505e"
/>