Create isActive property on cell alongside selectionStatus
The active state of a cell is orthogonal to its selectionStatus. A cell can be active and have a selection status of editing or selected. We need to track both states separately.

The isActive property on the cell is added out of convenience so the getActiveCell helper function doesn't need to be used in components that
already have access to the cell object.