catalog explorer: use connections.toml for snowflake (#10490)
This updates the Snowflake behavior for adding catalogs. It is now fully
driven by a `connections.toml` file; you can configure a path to a
custom location or will automatically look for envvar `$SNOWFLAKE_HOME`
(relevant for workbench) or `~/.snowflake` (or equivalent os paths).

This changes the behavior to _require_ users to have a
`connections.toml` file, which is idiomatic for snowflake users. This
gives much more flexibility for connection configuration without having
to manage this ourselves.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #10243 User's ROLE is visible in Catalog Explorer for Snowflake
connections.
- #10241 Catalog Explorer uses Snowflake's `connections.toml` to handle
connections.

#### Bug Fixes

- N/A


### QA Notes

things to check

- can add snowflake using a `connections.toml` file, all connections are
available for selection
- can expand nodes
- tooltip over includes ROLE if specified
- when added, the name repopulates in explorer ui on reload
	- click to connect
- can remove connection
	- when you reload, connection does not persist in explorer ui
- can run python code (note: requires driver setup)
- can run r code (note: requires driver setup)