Only show notebook cell context menu items in notebook editor (#10523)
Addresses #10518 

This PR addresses the issue where the notebook cell context menu
incorrectly appeared in the Output panel right-click menu WHEN a
positron notebook was the active editor.

I did some testing and noticed that the Output pane was the only place I
was seeing the notebook cell menu items. They don't show up if you
right-click in the Primary or Secondary Side Bars. It turns that this is
because the Output panel is a Monaco editor under the hood (based off a
quick look at the dom elements via dev tools) and contributes its
actions to `MenuId.EditorContext` which is the menu used by editors (see
`Clear Output`):

https://github.com/posit-dev/positron/blob/20c6627a71bdfaa5ddb880c16934a7cf408dca90/src/vs/workbench/contrib/output/browser/output.contribution.ts#L319.
This is the same menu that the notebook cell actions are contributed
too. Other editor specific actions such as "Go To Definition" which are
contributed to `MenuId.EditorContext` also check if the editor has focus
via context keys like `editorTextFocus` which prevents them from showing
up in the Output panel.

I think the solution here is to do what the other editor commands do,
and check if the notebook editor has focus.

**BEFORE - notebook cell menu in output panel when a notebook is the
active editor**


https://github.com/user-attachments/assets/e426209c-84de-4cc4-9ef5-b7a17a48c446


**AFTER - no notebook cell menu in output panel when a notebook is the
active editor**



https://github.com/user-attachments/assets/5ba56ea5-ede1-4a99-a070-46bc91df9a25



### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes
the main testing notes are in #10518 for the regression.

We should also test to make sure the submenu still appears in positron
notebooks as expected. I do want to point out that prior to this change,
the notebook cell context menu only showed up when a user right-clicked
in a cell when it was in edit mode this same behavior has been
maintained. You can't right click on a markdown cell that is NOT in edit
mode. I think this is because we contribute these menu items to the
"Editor" menu and the view mode of a markdown cell is not an editor so
the menu can't be brought up. I think this is something that we could
fix in the future but isn't part of this issue.

@:positron-notebooks

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->