Use URI when available for commands to run in console (R, Python) (#10590)
Addresses https://github.com/posit-dev/positron/issues/9989

The editor action bar has correctly been passing the resource URI of the
editor where the button was clicked, but the extension command handlers
were ignoring this parameter and instead using
`vscode.window.activeTextEditor`, which returns the focused editor, not
necessarily the one with the button. This PR updates the extension
commands to use the URI if it is passed in.

@:console
@:editor-action-bar
@:sessions

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fixed the "play" button behavior for split editors.


### QA Notes

If you set up Positron so that you have split your editor like this,
perhaps with a `.R` file on one side and a `.py` file on the other:

<img width="1428" height="890" alt="Screenshot 2025-11-15 at 11 30
36 AM"
src="https://github.com/user-attachments/assets/6d31edf1-5ee3-4cba-a88a-acc580f3cda0"
/>

you should be able to:

- put your cursor in the `.R` file and click the "play" button on the
`.py` file, and then see the Python file execute in the Python console
- similarly put your cursor in the `.py` file and click the "play"
button on the `.R` file, and then see the R file execute in the R
console