ci: max failure reduction on matrix runs (#10628)
### Summary
When leveraging matrix to run e2e tests (like in full suite), the max
failure was 10, but it was 10 _FOR EACH shard_ which in the case of a
run with a lot of failures ([upstream
merge](https://github.com/posit-dev/positron/actions/runs/19466823561)),
was creating massive blob reports for later merging and running out of
disk space. I updated max failures to be lower based on # of shards.
Also some small misc improvements.


### QA Notes
n/a