add back python.languageServer config (#10626)
Addresses https://github.com/posit-dev/positron/issues/10545 and
https://github.com/posit-dev/positron/issues/7393.

This PR adds back the `python.languageServer` setting as an enum with
one choice (None). This is because a handful of extensions rely on this
setting to exist in order to do things with it. See the linked issues
for examples.

I validated that you can now install pyrefly and basedpyright without
issues.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- The python LSP extensions `pyrefly` and `basedpyright` can now be
activated without popup errors.


### QA Notes

Try those extensions, and make sure notebooks behave the same as before
(due to a small change in how the jupyter extension might work)