Notebooks: Track the Active Cell For Selection (#10573)
Addresses #9848 #10280 #10281 #9992 which are all active
cell/focus/selection related bugs and #10470 which just updates the
clear output icon to match the icons we use elsewhere in the IDE.

This is a pretty large PR that addresses a number of selection behavior
bugs and some selection/hover styling requests as described in the
issues above. The root cause for many of these issues was that we didn't
really treat a single cell as the one with focus in multi-select
scenarios.

Following the commits to understand the changes would be an easy way to
understand what each change is doing.

**Fixed:**
* Updated "Clear Output" action bar button icon
* Cell action bar only shows up for active cell in single select and
multi-select scenarios
* The active cell now always shows the blue focus ring around the cell
to help differentiate it from a selected cell.
* Cell actions should now work in reference to the active cell
* Hovering over a cell that is not the active cell or a selected cell
will now have a grey border hover state. The cell action bar will NOT
show up when hovering over a cell since we only show the action bar for
the active cell.
* "+ Code" and "+ Markdown" action bar buttons now insert the cell after
the active cell and make it the new active cell
* You can escape a multi-select scenario by pressing `Escape`
* Multi-select scenarios now support de-selecting cells via
Shift+ArrowUp or Shift+ArrowDown. Previously the selection would always
grow the selection but now multi-selection has an "anchor" which is the
active cell. You can only increase a selection of cells in one
direction. Moving in the opposite direction will decrease the cell
selection. Ex: Given a file with 5 cells (cell 0 to cell 4), if the
active cell is cell 2, Shift+ArrowUp twice means selected cells are cell
0, cell 1, cell 2. Pressing Shift+ArrowDown twice makes the selection be
cell 2.

There are probably other things that may have been fixed with this
change but the above were intentionally fixed. I'd normally include
videos but there is A LOT that has changed and capturing it all in a few
videos isn't totally possible.

**Not Fixed:**
* Cut and delete operations for multi-select scenarios
  * This will be addressed in a follow up as part of #10116 
* Cmd+click to multi-select non-contiguous cells and Shift+click for
contiguous selection
* This has not been directly addressed but it does look cmd+click is
working for copying cells. Shift+click works the same as cmd+click which
is incorrect. All of these issues will be addressed in #10116

**Hover State**


https://github.com/user-attachments/assets/2bfb87f4-1667-4270-84fb-5282d8f6a822

**Basic Multi-Select + Active Cell Keyboard Navigation**


https://github.com/user-attachments/assets/66fe1f23-617e-4cbe-adaf-6fc0a912df0a


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes
@:positron-notebooks
<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->